requirements.txt
Flask==1.0.2
numpy>=1.20
Pillow==9.5.0
tensorflow==2.11.0
tensorflow_intel==0.0.1
gunicorn==19.9.0
Jinja2==2.11.
markupsafe==1.1.1
MarkupSafe>=2.1.1
itsdangerous==1.1.0
